[#28] refactor: 모달 합성 컴포넌트를 각각의 파일로 분리
