+++ src/components/Modal/index.ts
@@ -1,0 +1,8 @@
+export { default as Modal } from './Modal';
+export { default as ModalBody } from './ModalBody';
+export { default as ModalClose } from './ModalClose';
+export { default as ModalContent } from './ModalContent';
+export { default as ModalFooter } from './ModalFooter';
+export { default as ModalHeader } from './ModalHeader';
+export { default as ModalTitle } from './ModalTitle';
+export { default as ModalTrigger } from './ModalTrigger';
